moved idea post specs to separate file
