+++ back/spec/acceptance/ideas_create_spec.rb
@@ -1,0 +1,354 @@
+require 'rails_helper'
+require 'rspec_api_documentation/dsl'
+
+resource 'Ideas' do
+  explanation 'Proposals from citizens to the city.'
+
+  before do
+    header 'Content-Type', 'application/json'
+
+    create(:idea_status_proposed)
+  end
+
+  post 'web_api/v1/ideas' do
+    with_options scope: :idea do
+      parameter :project_id, 'The identifier of the project that hosts the idea', required: true
+      parameter :phase_ids, 'The phases the idea is part of, defaults to the current only, only allowed by admins'
+      parameter :author_id, 'The user id of the user owning the idea. This can only be specified by moderators and is inferred from the JWT token for residents.'
+      parameter :idea_status_id, 'The status of the idea, only allowed for admins', extra: "Defaults to status with code 'proposed'"
+      parameter :publication_status, 'Publication status', required: true, extra: "One of #{Post::PUBLICATION_STATUSES.join(',')}"
+      parameter :title_multiloc, 'Multi-locale field with the idea title', required: true, extra: 'Maximum 100 characters'
+      parameter :body_multiloc, 'Multi-locale field with the idea body', extra: 'Required if not draft'
+      parameter :topic_ids, 'Array of ids of the associated topics'
+      parameter :location_point_geojson, 'A GeoJSON point that situates the location the idea applies to'
+      parameter :location_description, 'A human readable description of the location the idea applies to'
+      parameter :proposed_budget, 'The budget needed to realize the idea, as proposed by the author'
+      parameter :budget, 'The budget needed to realize the idea, as determined by the city'
+      parameter :idea_images_attributes, 'an array of base64 images to create'
+      parameter :idea_files_attributes, 'an array of base64 files to create'
+      parameter :custom_field_name1, 'A value for one custom field'
+    end
+    ValidationErrorHelper.new.error_fields self, Idea
+    response_field :ideas_phases, "Array containing objects with signature { error: 'invalid' }", scope: :errors
+    response_field :base, "Array containing objects with signature { error: #{ParticipationPermissionsService::POSTING_DISABLED_REASONS.values.join(' | ')} }", scope: :errors
+
+    let(:idea) { build(:idea) }
+    let(:with_permissions) { false }
+    let(:project) { create(:single_phase_ideation_project, phase_attrs: { with_permissions: with_permissions }) }
+    let(:project_id) { project.id }
+    let(:publication_status) { 'published' }
+    let(:title_multiloc) { idea.title_multiloc }
+    let(:body_multiloc) { idea.body_multiloc }
+
+    context 'when visitor' do
+      describe 'default permissions' do
+        let(:author_id) { nil }
+
+        example '[error] Create an idea without author', document: false do
+          do_request
+          assert_status 401
+        end
+      end
+
+      describe "native survey response when permission is 'everyone'" do
+        before { IdeaStatus.create_defaults }
+
+        let(:project) do
+          create(:single_phase_native_survey_project, phase_attrs: { with_permissions: true }).tap do |project|
+            project.phases.first.permissions.find_by(action: 'posting_idea').update! permitted_by: 'everyone'
+          end
+        end
+        let(:project_id) { project.id }
+        let(:extra_field_name) { 'custom_field_name1' }
+        let(:form) { create(:custom_form, participation_context: project.phases.first) }
+        let!(:text_field) { create(:custom_field_text, key: extra_field_name, required: true, resource: form) }
+        let(:custom_field_name1) { 'test value' }
+
+        example_request 'Create a native survey response without author' do
+          assert_status 201
+          json_response = json_parse response_body
+          idea_from_db = Idea.find(json_response[:data][:id])
+          expect(idea_from_db.author_id).to be_nil
+          expect(idea_from_db.custom_field_values.to_h).to eq({
+            extra_field_name => 'test value'
+          })
+        end
+      end
+    end
+
+    context 'when resident' do
+      before { header_token_for(resident) }
+
+      with_options scope: :idea do
+        parameter :anonymous, 'Post this idea anonymously'
+      end
+
+      let(:resident) { create(:user) }
+      let(:topic_ids) { create_list(:topic, 2, projects: [project]).map(&:id) }
+      let(:location_point_geojson) { { type: 'Point', coordinates: [51.11520776293035, 3.921154106874878] } }
+      let(:location_description) { 'Stanley Road 4' }
+
+      describe 'creating an idea' do
+        example_request 'Create an idea' do
+          assert_status 201
+          expect(response_data.dig(:relationships, :project, :data, :id)).to eq project_id
+          expect(response_data.dig(:relationships, :topics, :data).pluck(:id)).to match_array topic_ids
+          expect(response_data.dig(:attributes, :location_point_geojson)).to eq location_point_geojson
+          expect(response_data.dig(:attributes, :location_description)).to eq location_description
+          expect(project.reload.ideas_count).to eq 1
+        end
+
+        example 'Check for the automatic creation of a like by the author when an idea is created', document: false do
+          do_request
+          assert_status 201
+          new_idea = Idea.find(response_data[:id])
+          expect(new_idea.reactions.size).to eq 1
+          expect(new_idea.reactions[0].mode).to eq 'up'
+          expect(new_idea.reactions[0].user.id).to eq resident.id
+          expect(response_data[:attributes][:likes_count]).to eq 1
+        end
+
+        describe 'Values for disabled fields are ignored' do
+          let(:proposed_budget) { 12_345 }
+
+          example 'Create an idea with values for disabled fields', document: false do
+            do_request
+            assert_status 201
+            expect(response_data.dig(:attributes, :title_multiloc, :en)).to eq 'Plant more trees'
+            # proposed_budget is disabled, so its given value was ignored.
+            expect(response_data.dig(:attributes, :proposed_budget)).to be_nil
+            expect(response_data.dig(:relationships, :topics, :data).pluck(:id)).to match_array topic_ids
+            expect(response_data.dig(:attributes, :location_point_geojson)).to eq location_point_geojson
+            expect(response_data.dig(:attributes, :location_description)).to eq location_description
+          end
+        end
+      end
+
+      describe 'when creating an idea that is a survey response' do
+        let(:project) { create(:single_phase_native_survey_project, default_assignee_id: create(:admin).id) }
+        let(:idea) { build(:native_survey_response, project: project) }
+
+        example 'does not assign anyone to the created idea', document: false do
+          do_request
+          assert_status 201
+          idea = Idea.find(json_parse(response_body).dig(:data, :id))
+          expect(idea.assignee_id).to be_nil
+          expect(idea.assigned_at).to be_nil
+        end
+
+        context 'creating a draft survey response' do
+          let(:publication_status) { 'draft' }
+
+          example 'sets the publication status to draft' do
+            do_request
+            assert_status 201
+            idea = Idea.find(json_parse(response_body).dig(:data, :id))
+            expect(idea.publication_status).to eq 'draft'
+          end
+        end
+      end
+
+      describe 'when posting an idea in an active ideation phase, the correct form is used' do
+        let(:project) { create(:project_with_active_ideation_phase) }
+        let!(:custom_form) { create(:custom_form, :with_default_fields, participation_context: project) }
+        let(:proposed_budget) { 1234 }
+
+        example 'Post an idea in an ideation phase' do
+          custom_form.custom_fields.find_by(code: 'proposed_budget').update!(enabled: true)
+
+          do_request
+
+          assert_status 201
+          idea = Idea.find(response_data[:id])
+          expect(idea.proposed_budget).to eq 1234
+        end
+      end
+
+      describe 'when posting an idea in an active ideation phase, the creation_phase is not set' do
+        let(:project) { create(:project_with_active_ideation_phase) }
+        let!(:custom_form) { create(:custom_form, participation_context: project) }
+
+        example_request 'Post an idea in an ideation phase' do
+          assert_status 201
+          json_response = json_parse response_body
+          idea = Idea.find(json_response.dig(:data, :id))
+          expect(idea.creation_phase).to be_nil
+        end
+      end
+
+      describe 'Creating an idea anonymously' do
+        let(:allow_anonymous_participation) { true }
+        let(:anonymous) { true }
+
+        before { project.phases.first.update! allow_anonymous_participation: allow_anonymous_participation }
+
+        example_request 'Posting an idea anonymously does not save an author id' do
+          assert_status 201
+          expect(response_data.dig(:attributes, :anonymous)).to be true
+          expect(response_data.dig(:attributes, :author_name)).to be_nil
+          expect(response_data.dig(:relationships, :author, :data)).to be_nil
+        end
+
+        example 'Does not add the author as a follower', document: false do
+          expect { do_request }.not_to change(Follower, :count)
+        end
+
+        example 'Does not log activities for the author', document: false do
+          expect { do_request }.not_to have_enqueued_job(LogActivityJob).with(anything, anything, resident, anything, anything)
+        end
+
+        describe 'when anonymous posting is not allowed' do
+          let(:allow_anonymous_participation) { false }
+
+          example_request 'Rejects the anonymous parameter' do
+            assert_status 422
+            json_response = json_parse response_body
+            expect(json_response).to include_response_error(:base, 'anonymous_participation_not_allowed')
+          end
+        end
+      end
+
+      describe 'Creating a native survey response when posting anonymously is enabled' do
+        let(:project) { create(:single_phase_native_survey_project, phase_attrs: { allow_anonymous_participation: true }) }
+
+        example_request 'Posting a survey automatically sets anonymous to true' do
+          assert_status 201
+          expect(response_data.dig(:attributes, :anonymous)).to be true
+          expect(response_data.dig(:attributes, :author_name)).to be_nil
+          expect(response_data.dig(:relationships, :author, :data)).to be_nil
+        end
+      end
+
+      describe 'Creating a native survey response when posting anonymously is not enabled' do
+        let(:project) { create(:single_phase_native_survey_project, phase_attrs: { allow_anonymous_participation: false }) }
+
+        example_request 'Posting a survey does not set the survey to anonymous' do
+          assert_status 201
+          expect(response_data.dig(:attributes, :anonymous)).to be false
+          expect(response_data.dig(:attributes, :author_name)).not_to be_nil
+          expect(response_data.dig(:relationships, :author, :data)).not_to be_nil
+        end
+      end
+
+      describe 'For projects without ideas_order' do
+        let(:project) { create(:single_phase_ideation_project, phase_attrs: { ideas_order: nil }) }
+
+        example 'Creates an idea', document: false do
+          do_request
+          assert_status 201
+          json_response = json_parse(response_body)
+          expect(json_response.dig(:data, :relationships, :project, :data, :id)).to eq project_id
+          expect(json_response.dig(:data, :relationships, :topics, :data).pluck(:id)).to match_array topic_ids
+          expect(json_response.dig(:data, :attributes, :location_point_geojson)).to eq location_point_geojson
+          expect(json_response.dig(:data, :attributes, :location_description)).to eq location_description
+          expect(project.reload.ideas_count).to eq 1
+        end
+      end
+
+      describe 'Errors' do
+        describe do
+          let(:publication_status) { 'fake_status' }
+
+          example_request '[error] Creating an invalid idea' do
+            assert_status 422
+            json_response = json_parse response_body
+            expect(json_response).to include_response_error(:publication_status, 'inclusion', value: 'fake_status')
+          end
+        end
+
+        describe do
+          let(:idea_image) { file_as_base64 'header.jpg', 'image/jpeg' }
+          let(:idea_images_attributes) { [{ image: idea_image }] }
+
+          example_request 'Create an idea with an image' do
+            assert_status 201
+            json_response = json_parse(response_body)
+            expect(json_response.dig(:data, :relationships, :idea_images)).to be_present
+          end
+        end
+
+        describe do
+          let(:idea_files_attributes) { [{ name: 'afvalkalender.pdf', file: encode_file_as_base64('afvalkalender.pdf') }] }
+
+          example_request 'Create an idea with a file' do
+            assert_status 201
+            json_response = json_parse(response_body)
+            expect(Idea.find(json_response.dig(:data, :id)).idea_files.size).to eq 1
+          end
+        end
+
+        describe do
+          let(:project) do
+            create(:project_with_current_phase, current_phase_attrs: {
+              participation_method: 'information'
+            })
+          end
+
+          example_request '[error] Creating an idea in a project with an active information phase' do
+            assert_status 401
+            json_response = json_parse(response_body)
+            expect(json_response.dig(:errors, :base).first[:error]).to eq 'not_ideation'
+          end
+        end
+
+        describe do
+          let(:project_id) { nil }
+
+          example_request '[error] Create an idea without a project' do
+            expect(response_status).to be >= 400
+          end
+        end
+
+        example '[error] Create an idea when there is a posting disabled reason' do
+          expect_any_instance_of(ParticipationPermissionsService)
+            .to receive(:posting_idea_disabled_reason_for_phase).with(project.phases.first, resident).and_return('i_dont_like_you')
+
+          do_request
+
+          assert_status 401
+          expect(json_parse(response_body)).to include_response_error(:base, 'i_dont_like_you')
+        end
+      end
+
+      example_group 'with permissions on phase' do
+        let(:with_permissions) { true }
+        let(:group) { create(:group) }
+
+        before do
+          project.phases.first.permissions.find_by(action: 'posting_idea')
+            .update!(permitted_by: 'groups', groups: [group])
+        end
+
+        example '[error] Create an idea in a project with groups posting permission', document: false do
+          do_request
+          assert_status 401
+        end
+
+        example 'Create an idea in a project with groups posting permission' do
+          group.add_member(resident).save!
+          do_request
+          assert_status 201
+        end
+      end
+
+      describe do
+        before { SettingsService.new.activate_feature! 'blocking_profanity' }
+
+        let(:title_multiloc) { { 'nl-BE' => 'Fuck' } }
+        let(:body_multiloc) { { 'fr-FR' => 'cocksucker' } }
+
+        example_request '[error] Create an idea with blocked words' do
+          assert_status 422
+          json_response = json_parse(response_body)
+          blocked_error = json_response.dig(:errors, :base)&.select { |err| err[:error] == 'includes_banned_words' }&.first
+          expect(blocked_error).to be_present
+          expect(blocked_error[:blocked_words].pluck(:attribute).uniq).to include('title_multiloc', 'body_multiloc')
+        end
+      end
+    end
+  end
+end
+
+def encode_file_as_base64(filename)
+  "data:application/pdf;base64,#{Base64.encode64(Rails.root.join('spec', 'fixtures', filename).read)}"
+end
